% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw04pchiyunliu-package.R
\docType{package}
\name{hw04pchiyunliu-package}
\alias{hw04pchiyunliu}
\alias{hw04pchiyunliu-package}
\title{hw04pchiyunliu: STAT 613 hw04 creating a package}
\description{
Create a package with the recursive sequence function for STAT613 hw04.
}
\author{
\strong{Maintainer}: ChiYun Liu \email{cl0471a@american.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
