% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myseq_n.R
\name{myseq_n}
\alias{myseq_n}
\title{Calculate the recursive sequence n
This is the recursive sequence. That is, element n is the sum of element n-1 and the difference between elements n-3 and n-2 divided by n.}
\usage{
myseq_n(x, n)
}
\arguments{
\item{x}{vector containing the first three numeric elements of the sequence}

\item{n}{element of the sequence}
}
\value{
element n
}
\description{
Calculate the recursive sequence n
This is the recursive sequence. That is, element n is the sum of element n-1 and the difference between elements n-3 and n-2 divided by n.
}
\examples{
myseq_n(x = c(2, 4, 3), n = 5)
}
